Add replacement FindPthreads
This patch fixes the pthread linking issue. Unfortunately, CMake only
considers the result from the C/C++ compiler and forces that flag be in
the interface link property. This means that the flag is transitively
passed to any libraries and executables. This is wrong for Swift since
it cannot take `-pthread`. For now, we'll just pass `-lpthread` (won't
work on Android, but will work for Linux for now).