Generate modulemaps from a directory
The flow targets have a list of headers attached to them in a property.
We pull that and turn them into a modulemap file.

At the moment, we're running into an issue where the generated header
files import the non-processed `*.actor.h` files cause the module to
fail to import. It would be better if those imported the actual C++
file.