Make parse errors of command-line knobs stop the program rather than emit a warning.
FDB binaries currently emit a warning to stderr as well as the logs when they
encounter command-line options for
* knobs which don't exist, and
* knob values which fail to parse.
They then proceed as if the command-line argument had never been provided.

FDB servers are often managed by automation, and no one sees warnings in the logs
or stderr unless they go looking for them. This means that a typo in the name
of a knob will result in a server running without the knob setting its operator
meant to set. Similarly, an attempt to specify a memory limit as a knob value
using command suffixes, e.g. '4GiB', will also fail, but the server will proceed
with the default value of that knob, contrary to the operator's intention.

I believe the safer option is to fail hard when we don't understand command-line
arguments, rather than emit a warning and proceed. Running with unexpecetd knob
values can lead to failures which are difficult to diagnose. Whereas failing
immediately at program start is straightforward to remedy.

New behavior:
$ bin/fdbserver --knob_commit_batches_mem_bytes_hard_limit=4GiB
ERROR: Invalid value '4GiB' for knob option 'commit_batches_mem_bytes_hard_limit'
Error: Option set with an invalid value