tuple: user defined type support (#8300)
* tuple: add support for usertype as last part of tuple

The expectation is user defined types come at the end as there is no
delimiter standard. In that case, if we encounter a user defined type,
we append it to the end of the tuple and allow users to transform that
type by parsing raw string returned.

* blob: use Tuple::unpackUserType() for key alignment

We want blob granule key alignment to be able to understand user data
types. This upgrades support and allows user data types to be the last
key in a blob granule and for us to be able to parse that properly.