Improve keyAfter, singleKeyRange, and an arena usage (#8674)
* Fix a potential memory error

The returned KeyRef should live at least as long as the supplied arena

* Improve keyAfter and singleKeyRange

Avoid duplicating the keyAfter implementation, and share memory between
begin and end for singleKeyRange returning a standalone

* Avoid creating and destroying an arena in a loop

This defeats some of the purposes of Arena's, namely to amortize the
cost of calling malloc and free and to improve cache locality.

* Improve Arena usage

Avoid an arena allocation in keyAfter - instead return a string with
static lifetime. I made sure to return the same memory as was just
brought into cache to inspect whether key == \xff\xff.

Also avoid creating and destroying an arena in a loop for encrypting
idempotency id sets.