Disallow uniform initialization syntax for state variables
The actor compiler does not handle this correctly, so disallow it for
now. We could probably support this someday if someone is sufficiently
motivated.

Previously this would miscompile and hopefully cause some kind of syntax
error in the generated code. Now it issues a helpful diagnostic.

Tested by adding an actor with a state variable that tries to use
uniform initialization syntax and observing the error message looks
good. We don't have any infrastructure for testing that the
ActorCompiler rejects something, so I'm not including a test in this
change. We could probably add that relatively easily though.