Disable Rollback fault injection workload for SpecialKeySpaceCorrectness test. (#8465)
One of the special keyspace tests reads values from
`\xff\xff/worker_interfaces' in two different ways. The rollback workload
can trigger a cluster recovery between the two reads. The cluster controller
can update the worker interface for a machine in that time, causing the
test to fail.

It's not expected that the worker_interfaces keyspace will remain unchanged
through a cluster reconfiguration. The purpose of the test was to check
that 'get' and 'getRange' returned consistent results for this special keyspace.

Therefore, we've disabled Rollback failures in this test.