apitester: add better support for blob api commands + tenant abstraction (#8700)
* blob/apitester: rename tenantDebugString() -> debugTenantStr()

* blob/apitester: fix int64_t initialization to -1

* blob/apitester: add missing override for randomOperation()

* blob/cpp: cpp blob tenant operations

* apitester: add tenant to execution context

* apitester: DatabaseOps abstraction

This creates a base abstract class for Database and Tenant operations to
share.

* blob/apitester: add setup phase to verifyRange() prior to operations

This let's us ensure blob storage is setup prior to running apitester
against the cluster.

* apitester: expand execOperation() to take a tenant

Now that db operations can be transparently applied to tenants or
databases through the IDatabaseOps interface, allow execOperation() to
take a tenant accordingly.

* blob/testerapi: blobbify in TesterApiWorkload setup

Before we were blobbifying wholesale through TestRunner. This makes the
change to blobbifying and verifying in setup, so the workload can
properly test database and tenant scope functions.