Add version indexer role.
This change introduces a new role, the version indexer, responsible for
maintaining the sequence of versions with data for each tag.

This is controllable via `configure version_indexerx=<n>`, which will
attempt to recruit `n` version indexer processes. By default, n=0, so
this feature is disabled.

Commit proxies will notify the version indexer of commits as they push
them to the logs. Storage servers can then poll the version indexer instead
of tlog for a definite sequence of versions since their last known version.

The benefits are two-fold.

The storage servers need only contact tlogs when they have data. They don't need to fetch empty versions.
Fewer tlogs are necessary, as this reduces the load on the tlogs (while offloading it to version indexers).
This is important because currently commit proxies must write to every tlog, to avoid a version gap in any
single tlog. Similarly, storage servers must read empty commits from tlogs. In the future, we hope to avoid
the empty tlog commits.

Note: There is a lot of boilerplate involved in adding a role, and the recruitment code
in the cluster controller is especially tricky. The meat of the change is a couple of small
updates to commit proxy and storage server, and of course the version indexer itself,
which is conceptually simple.

Note: By default, there are 0 version indexers, and the code should function as before.
This number can be changed by config, and of course we can set the default to 1
in the future.

Note: Despite version indexers being disabled by default, they are included in simulation
testing, so the code is well exercised.

Testing: several hundred thousand simulation tests, with no new failures attributed to the
version indexer addition.