test: add RAW_ACCESS when touching global special key space
When writing to the special key space, if RAW_ACCESS is not added, the
default tenant for the database is used. This causes a failure in the
SpecialKeySpace actor where it looks to see if accessing a key by tenant
is allowed, but if it isn't and there is a tenant, we fail. The fix is
RAW_ACCESS prevents propagation of the default tenant.