Track whether an AndFuture has begun (#8937)
* Track whether an AndFuture has begun

Previously, for deciding whether or not it was legal to set
READ_YOUR_WRITES_DISABLE, it was allowed as long as no reads were
in-flight. It's possible that a read can have been initiated, but the
RYW caches are empty and no reads are in-flight. See #8925 for an
example.

Closes #8925

* Change began to futureCount

* Remove unused and incorrect assignment operators

* Remove bogus test expectation

It shouldn't be legal to set READ_YOUR_WRITES_DISABLE after reading if
there are currently no pending reads, but this test was expecting that
to work.