add multi version api for prefetch
there are 2 cases we want to cover
1. A 7.1 FDB Java with a 7.2 FDB C library talking to a 7.2 fdbserver.
2. A 7.2 FDB Java with a 7.2 FDB C library using the multi-version
  client to talk to a 7.1 fdbserver via a second 7.1 FDB C library

For the first case, FDB_API_VERSION in FdbJNI.cpp is set in compile time,
and it will determine what interface it got from the C bindings,
as a result, it becomes 7.1 Java + 7.1 C talk to 7.2 server, and it
will reach fdb_transaction_get_mapped_range_v719 method.

For the second case, FDB_API_VERSION will be 7.2 and we check if runtime
API version is smaller than 720, if so fdb_transaction_get_mapped_range_719
will be called. Otherwise fdb_transaction_get_mapped_range_impl will be called.
However, in the former case, it requires a dummy value for matchIndex to be set,
thus I am not sure if we want to support this because java code will have this dummy
variable to be set in the signature, meanwhile setting it to 0 is the same behavior as
the old version in 7.1