Fix output for generated Unions
I believe the issue here is that using a vector of `std::string_view` to
store `std::string`s returned from a function is invalid. The stack
allocated `std::string`s go out of scope and the `std::string_view` no
longer refers to valid memory. The fix is to store a copy of the string,
at the cost of a bit of performance when generating the code.