add new method signature for optional index prefetch
it supports the case of old client and new server: in 7.2 C binding,
keep the interface from 7.1 C binding(it is needed by 7.1 java binding),
while introduce a new interface for the usage of 7.2 java or later.

Thus, 7.2 java always use 7.2 C binding, despite it still provides an old
interface to client. While 7.1 java always use 7.1 C binding, and 7.1 C
binding can be removed once all java on client is newer than 7.1.

As a result, when 7.1 java hits 7.2 C binding, it will find the old
interface in C binding and use it.