Add getMappedRangeV2 API with byte array in request/reply
Two more features are added in V2:
* optionally return index
* indicate whether a primary record is served by this SS taking prefetch index query

A new API is needed because the struct of req/rep have changed and is not backward
compatible.

To achieve this, Java takes parameters to indicate how to return index and
whether to do a server to server call in the case of a local miss, then
encode them in a byte array. C API takes a byte array and knows how to parse it.
The reason to have byte array is for extendability -- new fields can be encoded
in the byte array without a signature/protocol change.

It is easy to add new fields to request or reply of getMappedRangeV2 API:
instead of creating another API, we only need to bump the API-defined protocol version,
as long as the client and server both know how to parse the byte array in req/rep.

QUICK_GET_KEY_VALUES_FALLBACK is removed, old API will always try to do server
to server call to fetch KV from other SS. While new API can specify this parameter.