BUGGIFY `safeThreadFutureToFuture()`  (#9263)
The `safeThreadFutureToFuture` function converts a `ThreadFuture` to a Future. There are a few cases where we have a
 result of type `ThreadFuture<Standalone<T>>` or `ThreadFuture<Optional<Standalone<T>>` where the memory is not
actually owned by the Standalone. Rather it is owned by the ThreadFuture.

Eventually we should fix this so that the memory is properly owned by the `Standalone`, which is beyond the scope of
this PR. Until then, we should update the implementation of `safeThreadFutureToFuture` to take away ownership from the
`Standalone`/`Optional<Standalone>` so we can detect this problem in simulation.

Test plan:
- correctness test with and without valgrind