Provide a tool that allows downloading logs when simulation RocksDB failures
A script, rocksdb_logtool.py, is available to upload/download generated
XML/JSON log files when test harness 2 detects that the test is failed,
and the script detects that the test is using RocksDB.

To upload, no actions is needed, using the regular

```
joshua start --tarball correctness.tgz
```

The build system will automatically pack the rocksdb_logtool.py into the
tarball and test harness 2 will call the script if it thinks the test is
failed.

To download, simply provide the ensemble id and test uid, e.g.

```
python3 rocksdb_logtool.py download --ensemble-id
20230222-204240-xiaogesu-cb6ea277a898f134 --test-uid
ab6fb792-088f-49d6-92d2-43bc4fb81668
```

Note the test UID can be retrieved by

```
joshua tail ensemble_id
```

output, it is in the field `TestUID` in <Test> element of test harness 2
generated XML.

For convenience, it is possible to do a

```
python3 rocksdb_logtool.py list --ensemble-id ensemble-id
```

to generate all possible download commands for failed tests. However,
the list subcommand will *NOT* verify if the test failure is coming from
RocksDB, i.e. other test failues may be included and it is the caller's
responsibility to verify. If the test is not RocksDB related, the
download would fail as nothing is uploaded.