Fix overflow/underflow in Histogram.sampleSeconds. (#9548)
The conversion from double -> uint32_t in sampleSeconds is potentially
undefined behavior for very large or negative values. We safely clamp
the value to [0, UINT32_MAX] before converting to integer.

Note: negative values are possible here, because timers can go backwards.
That case should be clamped to zero rather than underflowing.

This problem was discovered because we were mixing g_network->timer() and
timer(). Normally, these are the same and comparable. In simulation, they
are not and subtracting them led to the undefined behavior above. These
sites are now fixed.