Replace 10-second delay with explicit wait for cluster recovery in checkExtraDataStores.
CheckExtraDataStores reboots or kills storage servers with extra data stores.
Since this occurs during a consistency check, the expectation is that the database
is quiet and not in the midst of recovery. This was done with a 10-second delay,
but it's possible during simulation tests that it takes longer than 10 seconds
to recruit a new master, so this assumption is invalid and can cause a test failure
when the consistency checks proceed.

Instead of a delay, we run an empty transaction through the system and explicitly
wait for the cluster to return to a fully-recovered state.