A script that helps cherry picking PR commits to other branches
Usage:
    cherry-pick-upstream.py [PR id] [PR id] ...

Run the script in the branch that will receive the PRs.

It handles merge/fast-forward pull requests. If the commits are already
in the current branch, the script will do nothing.

The author thanks Henry Phan (hphan3@apple.com) for helpful discussion
about the relationship between Pull Requests and merge commits.