Revert error masking in #9913
So that it wouldn't hide bugs where something in the disk queue or storage
engine gets stuck but works after a reboot as the test wouldn't fail so the
hang would not be exposed as an issue.

The issue for #9913 is solved by increasing TLOG_MAX_CREATE_DURATION in
simulation, which is already done in commit 0aa096dc1.