Move recruitment functions to Recruiter class
Most functions don't need to be public, and have been moved to an
anonymous namespace in the Recruiter implementation. The public
interface for recruitment has been greatly narrowed to just a few
functions. With further improvements we should be able to narrow it even
more.

There is follow up work to reduce the dependence of recruitment
functions on `ClusterControllerData`.